Revert "Revert "chore(backend): Moved shared module to util/shared and drop unused""
This reverts commit 0df7e2cb08a94f9a13fef4283c762af08605ad23.
